Move email models into a folder
