+++ src/Umbraco.Core/Mail/IEmailSender.cs
@@ -1,5 +1,5 @@
 using System.Threading.Tasks;
-using Umbraco.Cms.Core.Models;
+using Umbraco.Cms.Core.Models.Email;
 
 namespace Umbraco.Cms.Core.Mail
 {
