+++ src/Umbraco.Core/Mail/NotImplementedEmailSender.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Threading.Tasks;
-using Umbraco.Cms.Core.Models;
+using Umbraco.Cms.Core.Models.Email;
 
 namespace Umbraco.Cms.Core.Mail
 {
