+++ src/Umbraco.Core/Notifications/SendEmailNotification.cs
@@ -1,4 +1,4 @@
-using Umbraco.Cms.Core.Models;
+using Umbraco.Cms.Core.Models.Email;
 
 namespace Umbraco.Cms.Core.Notifications
 {
