+++ src/Umbraco.Web.BackOffice/Controllers/AuthenticationController.cs
@@ -17,6 +17,7 @@ using Umbraco.Cms.Core.Mail;
 using Umbraco.Cms.Core.Mapping;
 using Umbraco.Cms.Core.Models;
 using Umbraco.Cms.Core.Models.ContentEditing;
+using Umbraco.Cms.Core.Models.Email;
 using Umbraco.Cms.Core.Models.Membership;
 using Umbraco.Cms.Core.Net;
 using Umbraco.Cms.Core.Security;
